use relative imports for mixpanel code
