+++ lib/metrics/mixpanel/interfaces/PageEvent.ts
@@ -1,5 +1,30 @@
-import type { BaseEvent } from 'lib/metrics/mixpanel/interfaces/BaseEvent';
+import type { PageType } from '@prisma/client';
+
+import type { BaseEventWithoutGroup, BaseEvent } from './BaseEvent';
 
 export type PageEvent = BaseEvent & {
   pageId: string;
 };
+
+type CustomPageType = 'settings' | 'proposals_list' | 'bounties_list' | 'nexus' | 'profile' | 'integrations';
+
+type CustomPageViewEvent = BaseEventWithoutGroup & {
+  pageId?: string;
+  spaceId?: string;
+  type: CustomPageType;
+};
+
+type PageTypeEvent = PageEvent & {
+  type: PageType;
+};
+
+export interface PageEventMap {
+  page_view: PageTypeEvent | CustomPageViewEvent;
+  archive_page: PageEvent;
+  delete_page: PageEvent;
+  restore_page: PageEvent;
+  edit_page: PageEvent;
+  create_page: PageTypeEvent;
+}
+
+export type PageEventNames = keyof PageEventMap;
