+++ lib/metrics/mixpanel/interfaces/ProposalEvent.ts
