+++ lib/metrics/mixpanel/interfaces/SettingEvent.ts
