+++ lib/metrics/mixpanel/interfaces/index.ts
@@ -1,9 +1,8 @@
-import type { BountyEventMap } from 'lib/metrics/mixpanel/interfaces/BountyEvents';
-import type { PageEventMap } from 'lib/metrics/mixpanel/interfaces/PageEvents';
-import type { SettingEventMap } from 'lib/metrics/mixpanel/interfaces/SettingEvents';
-
-import type { ProposalEventMap } from './ProposalEvents';
-import type { UserEventMap } from './UserEvents';
+import type { BountyEventMap } from './BountyEvent';
+import type { PageEventMap } from './PageEvent';
+import type { ProposalEventMap } from './ProposalEvent';
+import type { SettingEventMap } from './SettingEvent';
+import type { UserEventMap } from './UserEvent';
 
 export interface MixpanelTrackBase {
   // distinct_id - property name required by mixpanel to identify unique users
