+++ lib/metrics/mixpanel/updateTrackUserProfileById.ts
@@ -1,5 +1,6 @@
-import { updateTrackUserProfile } from 'lib/metrics/mixpanel/updateTrackUserProfile';
 import { getUserProfile } from 'lib/users/getUser';
+
+import { updateTrackUserProfile } from './updateTrackUserProfile';
 
 export async function updateTrackUserProfileById(userId: string) {
   try {
