+++ lib/users/createUser.ts
@@ -3,7 +3,7 @@ import { v4 } from 'uuid';
 
 import { prisma } from 'db';
 import getENSName from 'lib/blockchain/getENSName';
-import type { SignupAnalytics } from 'lib/metrics/mixpanel/interfaces/UserEvents';
+import type { SignupAnalytics } from 'lib/metrics/mixpanel/interfaces/UserEvent';
 import { trackUserAction } from 'lib/metrics/mixpanel/trackUserAction';
 import { updateTrackUserProfile } from 'lib/metrics/mixpanel/updateTrackUserProfile';
 import { isProfilePathAvailable } from 'lib/profile/isProfilePathAvailable';
